Add new ISearchEntity interface, implement dataset plugin
The ISearchEntity encapsulates all logic related to the indexing and
querying of a particular entity type, either core ones like datasets or
orgs or custom ones defined by the user.

It contains methods to return the search schema to use for this entity
(i.e. indexed fields) and a search query schema to add additional query
params that are entity-specific.

In addition contains methods to retrieve the entities to index, and a
list of all the entity ids.

This includes a first draft of the ISearchEntity plugin for datasets,
still not hooked to the rest of the logic.